+++ Chess.Domain/Models/Moves/CpuMoveCheckMate.cs
@@ -2,7 +2,7 @@ namespace Chess.Domain.Models.Moves
 {
     public class CpuMoveCheckMate : ICpuMoveResponse
     {
-        public int Value => -99;
+        public decimal Value => -99;
 
         public override string ToString() => "MATE";
     }
