+++ Chess.Domain/Models/Moves/CpuMoveUnknown.cs
@@ -2,7 +2,7 @@ namespace Chess.Domain.Models.Moves
 {
     public class CpuMoveUnknown : ICpuMoveResponse
     {
-        public int Value => 0;
+        public decimal Value => 0;
 
         public override string ToString() => "...";
     }
