+++ Chess.Domain/Models/Moves/ICpuMoveResponse.cs
@@ -2,6 +2,6 @@ namespace Chess.Domain.Models.Moves
 {
     public interface ICpuMoveResponse
     {
-        int Value { get; }
+        decimal Value { get; }
     }
 }
